+++ tests/ReloadTest.php
@@ -1,0 +1,14 @@
+<?php
+namespace Superhelio\Commands;
+
+use Superhelio\Commands\Commands\Reload;
+use Illuminate\Foundation\Testing\DatabaseMigrations;
+use Illuminate\Foundation\Testing\DatabaseTransactions;
+
+class ReloadTest extends \PHPUnit_Framework_TestCase
+{
+    public function testReloadTest()
+    {
+        $this->assertTrue(true);
+    }
+}