+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.0.1 - 2017-02-07
-### Added
+## [1.1.0](https://github.com/superhelio/commands/tree/2.0.0) (2017-08-27)
+- New command: `superhelio:gozer` with tests
+
+## [1.0.0](https://github.com/superhelio/commands/tree/1.0.0) (2017-02-08)
+- Fixed Scrutinizer reported issue
+- Renamed LICENSE
+- Tested `github_changelog_generator`, but did not like it
+
+## [0.0.1](https://github.com/superhelio/commands/tree/0.0.1) (2017-02-07)
 - `superhelio:reload`